Add Automerge integration
This commit adds collaboration using Automerge. The approach here is to
use the automerge document as a simple log of the grove patches, details
on the reasoning behind this are in the README.md.

In order to make this work accross devices I had to change the ID
generation to use UUIDs. The changes to `src/grove.ts` should be 
fairly mechanical. I also had to modify the `initial_client_state`
function to return the patches used to create the state as well as the
state itself so that those patches could be written to the automerge
document.